Fixed potential bug that could provide incorrect rebuilding from database (sliceDuration is now in doube precision)

+++ fr.inria.soctrace.tools.ocelotl.core/src/fr/inria/soctrace/tools/ocelotl/core/timeslice/TimeSliceVariableManager.java
@@ -15,9 +15,9 @@ public class TimeSliceVariableManager extends TimeSliceStateManager{
 			final TimeRegion testedTimeRegion, double value) {
 		final Map<Long, Double> timeSlicesDistribution = new HashMap<Long, Double>();
 		long startSlice = Math.max(
-				0,
-				(testedTimeRegion.getTimeStampStart() - timeRegion
-						.getTimeStampStart()) / sliceDuration - 1);
+				0,(long)
+				((testedTimeRegion.getTimeStampStart() - timeRegion
+						.getTimeStampStart()) / sliceDuration) - 1);
 		double temp = 0;
 		if (testedTimeRegion.getTimeStampStart()
 				- timeRegion.getTimeStampStart() >= 0)
